tu.SkipIssue mechanism
